adding http call for feedback

+++ src/app/services/loading-bar-interceptor.service.ts
@@ -9,13 +9,23 @@ export class LoadingBarInterceptor implements HttpInterceptor {
   constructor(private loadingBarService: LoadingBarService) {}
 
   intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+
+    if(req.headers.get('disableLoadingBar')){
+      return next.handle(req);
+    }
+
     this.loadingBarService.addRequest();
 
     const handleObs: Observable<HttpEvent<any>> = next.handle(req);
 
-    return handleObs.pipe(tap(()=> {
-      console.log('b');
+    const obs = handleObs.pipe(tap(()=> {
+      this.loadingBarService.removeRequest();
+    }, ()=>{
       this.loadingBarService.removeRequest();
     }));
+
+    this.loadingBarService.setRequestObs(obs);
+
+    return obs;
   }
 }
